[FLINK-34440] Incorporate CR comment

+++ flink-formats/flink-protobuf-confluent-registry/src/main/java/org/apache/flink/formats/protobuf/registry/confluent/ProtoToRowDataConverters.java
@@ -553,7 +553,7 @@ public class ProtoToRowDataConverters {
             return new ProtoToRowDataConverter() {
                 @Override
                 public Object convert(Object object) {
-                    return ((Number) extractValueField(object)).byteValue();
+                    return ((Number) extractValueField(object)).longValue();
                 }
             };
         } else {
